Restructure package, clean up some validator/paginator args, fix tests (all work now !) stub out new docs
