+++ aiotnb/errors.py
