+++ aiotnb/schemas.py
@@ -1,0 +1,114 @@
+"""
+The MIT License (MIT)
+
+Copyright (c) 2021 AnonymousDapper
+"""
+
+from __future__ import annotations
+
+__all__ = (
+    "partial",
+    "Key",
+    "AccountNumber",
+    "BalanceLock",
+    "Timestamp",
+    "Signature",
+    "Url",
+    "BankConfig",
+    "AccountSchema",
+)
+
+from datetime import datetime
+from typing import TYPE_CHECKING
+
+from nacl.encoding import HexEncoder
+from nacl.signing import VerifyKey
+from yarl import URL
+
+from .enums import NodeType, UrlProtocol
+from .validation import As, Const, Fn, Ignore, Maybe, Schema, Type
+
+if TYPE_CHECKING:
+    from typing import Callable, Optional, TypeVar
+
+    R = TypeVar("R")
+
+# as soon as we have proper ParamSpec support, delete this mess
+def partial(fn: Callable[..., R], *args: ..., **kwargs: ...) -> Callable[..., R]:
+    def inner(*a: ..., **k: ...) -> R:
+        return fn(*args, *a, **kwargs, **k)
+
+    return inner
+
+
+def _parse_iso8601_utc(timestamp: str) -> datetime:
+    """
+    Attempts to parse an ISO8601 timestamp.
+    """
+
+    if timestamp[-1] == "Z":
+        timestamp = f"{timestamp[:-1]}+00:00"
+
+    return datetime.fromisoformat(timestamp)
+
+
+def _key_from_str(key_str: str) -> VerifyKey:
+    return VerifyKey(key_str.encode("utf-8"), encoder=HexEncoder)
+
+
+def _to_bytes(data: str, *, exact_len: Optional[int]) -> bytes:
+    if exact_len is not None and len(data) != exact_len:
+        raise ValueError(f"value should be {exact_len} bytes")
+
+    return data.encode("utf-8")
+
+
+# Schema models start here
+
+Key = partial(As, str)
+
+AccountNumber = Key(_key_from_str)
+
+BalanceLock = Key(_to_bytes)
+
+Timestamp = Key(_parse_iso8601_utc)
+
+Signature = Key(partial(_to_bytes, exact_len=128))
+
+Url = Key(URL)
+
+
+# Main schemas
+
+BankConfig = Schema(
+    {
+        "primary_validator": dict,
+        "account_number": AccountNumber,
+        "ip_address": Url,
+        "node_identifier": AccountNumber,
+        "port": Maybe(int),
+        "protocol": Key(UrlProtocol),
+        "version": str,
+        "default_transaction_fee": int,
+        "node_type": Key(NodeType),
+    }
+)
+
+
+AccountSchema = {
+    "id": str,
+    "created_date": Timestamp,
+    "modified_date": Timestamp,
+    "account_number": AccountNumber,
+    "trust": Key(float),
+}
+
+
+# Helpers
+
+PAGINATOR_BASE = {
+    "count": int,
+    "next": Maybe(Url),
+    "previous": Maybe(Url),
+    "results": None,
+}
